+++ src/pages/interact-contract/index.tsx
@@ -1,0 +1,3 @@
+import InteractContract from "lib/pages/interact-contract";
+
+export default InteractContract;